refactor(maps): extract useMapSelectionModes from useMapFeaturesDraw
Move the four selecting-mode handlers behind one applySelectingMode(), collapsing endDraw's two selectingMode ladders.
